Moved fetch_events
